Add files via upload
This pcb/circuit is designed to use a Wemos D1 mini chip which I chose because they are fast, and I have a bunch of them (inexpensive as well) You can modify code and diagram as you want. I am not an engineer so there will probably better ways to make circuits. I just wanted to make something that uses the existing points system only as an input to tell the chip to control a high current NPN mosfet to cycle the ground side (-) of the ignition coil.